% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModAnnot.R
\name{AdaptedIdeogramTrackWithoutBandsData}
\alias{AdaptedIdeogramTrackWithoutBandsData}
\title{AdaptedIdeogramTrackWithoutBandsData Function (ModAnnot)}
\usage{
AdaptedIdeogramTrackWithoutBandsData(
  grangesGenome,
  cContigToViz,
  cOrgAssemblyName
)
}
\arguments{
\item{grangesGenome}{A GRanges object containing the width of each contig.}

\item{cContigToViz}{A character vector describing which contigs this track list should be prepared for.}

\item{cOrgAssemblyName}{The name of the genome assembly provided.}
}
\description{
Subfunction to return an empty IdeogramTrack without bands data.
}
\keyword{internal}
