% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloModAn.R
\name{addLetter}
\alias{addLetter}
\title{addLetter Function from seqLogo package (GloModAn)}
\usage{
addLetter(letters, which, x.pos, y.pos, ht, wt, fill)
}
\description{
Internal function from seqLogo package. See seqLogo documentation for more details.
User should use the addLetter function from seqLogo and should not use directly this version.
}
\keyword{internal}
