% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloModAn.R
\name{.AddModMotifPctToDf}
\alias{.AddModMotifPctToDf}
\title{AddModMotifPctToDf Function (GloModAn)}
\usage{
.AddModMotifPctToDf(
  dReportTable,
  grangesModPos,
  grangesGenome,
  dnastringsetGenome,
  cBaseLetterForMod,
  cModNameInOutput,
  nModPositionInMotif = 1,
  nUpstreamBpToAdd = 0,
  nDownstreamBpToAdd = 0
)
}
\arguments{
\item{dReportTable}{The dataframe in which the motif percentages must be included.}

\item{grangesModPos}{A GRanges object containing Modifications Positions data.}

\item{grangesGenome}{A GRanges object containing the width of each contig.}

\item{dnastringsetGenome}{A DNAStringSet object containing the sequence for each contig.}

\item{cBaseLetterForMod}{The name of the base letter of the modified base.}

\item{cModNameInOutput}{Name for the modification in the output.}

\item{nModPositionInMotif}{Number representing the position of the modification in the ranges
(after resizing if some base pairs were added). Defaults to 1.}

\item{nUpstreamBpToAdd}{Number of base pairs to add upstream of the range from the GRanges object provided
to obtain some sequence upstream of range. If some new ranges do not fit in the ranges of the contigs (provided with grangesGenome),
those new ranges will be removed. New windows with gaps are also removed. Defaults to 0.}

\item{nDownstreamBpToAdd}{Number of base pairs to add downstream of the range from the GRanges object provided
to obtain some sequence downstream of range. If some new ranges do not fit in the ranges of the contigs (provided with grangesGenome),
those new ranges will be removed. New windows with gaps are also removed. Defaults to 0.}
}
\description{
Return the dataframe provided with the motif percentages associated to modifications in the GRanges object.
}
\keyword{internal}
