% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeQual.R
\name{.CalculateNLx}
\alias{.CalculateNLx}
\title{CalculateNLx Function (SeQual)}
\usage{
.CalculateNLx(dnastringsetGenome, nNLx = 50)
}
\arguments{
\item{dnastringsetGenome}{A DNAStringSet object containing the sequence for each contig.}

\item{nNLx}{A number for the Nx and Lx to calculate. Defaults to 50 (so N50 and L50 calculation).}
}
\description{
Calculate Nx and Lx.
}
\keyword{internal}
