% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDREst.R
\name{.DrawFdrEst}
\alias{.DrawFdrEst}
\title{DrawFdrEst Function (FDREst)}
\usage{
.DrawFdrEst(
  dFdr,
  cNameParamToTest,
  nFdrPropForFilt,
  cMotifFg = NULL,
  lAdjustFdr = TRUE
)
}
\arguments{
\item{dFdr}{A dataframe with 3 columns:
\itemize{
  \item fdr: The false discovery rate estimated for this threshold.
  \item threshold: The threshold on the parameter.
  \item fdr_cummin: The minimum false discovery rate estimated for this threshold and less stringent thresholds (adjusted false discovery rate).
}}

\item{cNameParamToTest}{The name of the column containing the parameter to be tested.}

\item{nFdrPropForFilt}{A number indicating the false discovery rate to be used for filtering: this will allow to choose the closest threshold
below this number and represent it on the plot. Defaults to 0.05 (so fdr of 5\%).}

\item{cMotifFg}{The name of the motif tested. If NULL, the title will display "(Sample vs Control)" instead. Defaults to NULL.}

\item{lAdjustFdr}{If TRUE, display fdr_cummin (adjusted false discovery rate)
instead of fdr column. Defaults to TRUE.}
}
\description{
Return a plot describing the false discovery rate estimations by threshold on the parameter provided for the dataframe provided.
}
\keyword{internal}
