% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDREst.R
\name{.DrawFdrSegmentsXYAxis}
\alias{.DrawFdrSegmentsXYAxis}
\title{DrawFdrSegmentsXYAxis Function (FDREst)}
\usage{
.DrawFdrSegmentsXYAxis(dFdr, nFdrPropForFilt)
}
\arguments{
\item{dFdr}{A dataframe with 3 columns:
\itemize{
  \item fdr: The false discovery rate estimated for this threshold.
  \item threshold: The threshold on the parameter.
  \item fdr_cummin: The minimum false discovery rate estimated for this threshold and less stringent thresholds (adjusted false discovery rate).
}}

\item{nFdrPropForFilt}{A number indicating the false discovery rate to be used for filtering: this will allow to choose the closest threshold
below this number and represent it on the plot. Defaults to 0.05 (so fdr of 5\%).}
}
\description{
Draw a line describing the threshold associated to a value of false discovery rate estimation (defined by nFdrPropForFilt).
}
\keyword{internal}
