% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDREst.R
\name{.EstimateFdr}
\alias{.EstimateFdr}
\title{EstimateFdr Function (FDREst)}
\usage{
.EstimateFdr(nFg, nBg, nThreshold)
}
\arguments{
\item{nFg}{A numeric vector considered as the foreground signal.}

\item{nBg}{A numeric vector considered as the background signal.}

\item{nThreshold}{A numeric vector describing the thresholds used.}
}
\description{
Return a dataframe with the false discovery rate estimated by threshold on the parameter provided with the GRanges object(s).
}
\keyword{internal}
