% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModAnnot.R
\name{.ExportBamForGviz}
\alias{.ExportBamForGviz}
\title{.ExportBamForGviz Function (ModAnnot)}
\usage{
.ExportBamForGviz(
  cFileName,
  grangesObject,
  cBamXaParameter,
  dnastringsetGenome
)
}
\arguments{
\item{cFileName}{Name of the file to be exported.}

\item{grangesObject}{A GRanges object to be exported.}

\item{cBamXaParameter}{The name of the parameter to be exported as a "Xa" optional field within the bam file.}

\item{dnastringsetGenome}{A DNAStringSet object containing the sequence for each contig.}
}
\description{
Subfunction to export a GRanges object as a bam file.
}
\keyword{internal}
