% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModAnnot.R
\name{.ExportBigwigForGviz}
\alias{.ExportBigwigForGviz}
\title{.ExportBigwigForGviz Function (ModAnnot)}
\usage{
.ExportBigwigForGviz(
  cFileName,
  grangesObject,
  cBigwigParameter,
  lBigwigParametersByStrand,
  dnastringsetGenome
)
}
\arguments{
\item{cFileName}{Name of the file to be exported.}

\item{grangesObject}{A GRanges object to be exported.}

\item{cBigwigParameter}{The name of the parameter to be exported within the bigwig file.}

\item{lBigwigParametersByStrand}{If TRUE, the parameter is negative if associated range is on the reverse strand.}

\item{dnastringsetGenome}{A DNAStringSet object containing the sequence for each contig.}
}
\description{
Subfunction to export a GRanges object as a bw (bigwig) file.
}
\keyword{internal}
