% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModAnnot.R
\name{.ExportFastaForGviz}
\alias{.ExportFastaForGviz}
\title{.ExportFastaForGviz Function (ModAnnot)}
\usage{
.ExportFastaForGviz(cFileName, dnastringsetObject)
}
\arguments{
\item{cFileName}{Name of the file to be exported.}

\item{dnastringsetObject}{A DNAStringSet object to be exported.}
}
\description{
Subfunction to export DNAStringset object as fasta file.
}
\keyword{internal}
