% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModAnnot.R
\name{.GetCountsByDist}
\alias{.GetCountsByDist}
\title{GetCountsByDist Function (ModAnnot)}
\usage{
.GetCountsByDist(grangesDist, lAddCorrectedDistFrom5pTo3p = TRUE)
}
\arguments{
\item{grangesDist}{A GRanges objects containing ranges of given "Positions"
with their distance to feature positions.}

\item{lAddCorrectedDistFrom5pTo3p}{If TRUE, the distance will be corrected to reflect 5' to 3' direction
and will be stored in a new column (dist_5to3). Defaults to TRUE.}
}
\value{
A table objects containing "Mod" (or "Base") counts by distance to feature positions.
If a "Mod" (or "Base") is within nWindowSizeAroundFeaturePos base pairs of x different feature
positions: this "Mod" (or "Base")
will then reported x times with the distance to each feature position.
}
\description{
Return, in a table, the counts (or proportion) of given "Positions" by distance from feature
positions.
}
\keyword{internal}
