% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDREst.R
\name{.GetCumsumTableParamByMotif}
\alias{.GetCumsumTableParamByMotif}
\title{GetCumsumTableParamByMotif Function (FDREst)}
\usage{
.GetCumsumTableParamByMotif(
  grangesDataWithSeq,
  cNameParamToTest,
  nRoundDigits = 1,
  lTestBetweenMotifs = FALSE
)
}
\arguments{
\item{grangesDataWithSeq}{A GRanges-like object containing, in the extra columns, the parameter to be tested and the sequence.}

\item{cNameParamToTest}{The name of the column containing the parameter to be tested.}

\item{nRoundDigits}{The number of digits for the thresholds on the parameter. A value of 0 would mean no value to be rounded:
this could results in errors for a continuous variable. Defaults to 1.}

\item{lTestBetweenMotifs}{If TRUE, add one column for each sequence in the GRanges object provided.
Else, calculates the cumulative number of values of the parameter provided independantly of the sequence associated. Defaults to FALSE.}
}
\description{
Return a dataframe with cumulative number of values of the parameter provided in the GRanges object (from highest to lowest values).
}
\keyword{internal}
