% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModAnnot.R
\name{.GetModBaseCountsCategories}
\alias{.GetModBaseCountsCategories}
\title{GetModBaseCountsCategories Function (ModAnnot)}
\usage{
.GetModBaseCountsCategories(grangesAnnotationsWithCounts)
}
\arguments{
\item{grangesAnnotationsWithCounts}{A GRanges object based on grangesAnnotations with the counts:
\itemize{
  \item Modcount:           The number of "Mod"  within this feature.
  \item Modcount_perkbp:   (The number of "Mod"  within this feature divided by the size of the feature)*1000.
  \item Basecount:          The number of "Base" within this feature.
  \item Basecount_perkbp:  (The number of "Base" within this feature divided by the size of the feature)*1000.
}
The Genomic features categories must be in a column named "type".}
}
\description{
Subfunction to return categories of counts (or counts per KiloBase pairs (kbp)) of
the base modified (Mod) and the base letter of the modified base (Base). These categories
should have a comparable amount of Mod (or Base) between them.
Example: for Mod="6mA", Base="A"; for Mod="5mC", Base="C".
}
\keyword{internal}
