% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModAnnot.R
\name{.GetPosDistFromFeaturePos}
\alias{.GetPosDistFromFeaturePos}
\title{.GetPosDistFromFeaturePos Function (ModAnnot)}
\usage{
.GetPosDistFromFeaturePos(
  grangesData,
  grangesAnnotationsPos,
  cWhichStrandVsFeaturePos,
  nWindowSizeAroundFeaturePos,
  lAddCorrectedDistFrom5pTo3p
)
}
\arguments{
\item{grangesData}{A GRanges object containing the Mod (or Base) positions to be compared with
the grangesAnnotationsPos positions.}

\item{grangesAnnotationsPos}{A GRanges object containing the annotation positions to be compared with
the grangesData positions. The Genomic features categories must be in a column named "type".}

\item{cWhichStrandVsFeaturePos}{A character value describing if distance comparison must be made between "Mod"
(or "Base") and the feature positions...
\itemize{
  \item "same":    ...if these positions are on the same strand  only.
  \item "opposite":...if these positions are on opposite strands only.
  \item "both":    ...for all of these positions: same and opposite strands.
}}

\item{nWindowSizeAroundFeaturePos}{Size, in base pairs, of the viewing window around the feature positions.}

\item{lAddCorrectedDistFrom5pTo3p}{If TRUE, the distance will be corrected to reflect 5' to 3' direction
and will be stored in a new column (dist_5to3). Defaults to TRUE.}
}
\description{
Return, in a GRanges object, the distance between "Mod" (or "Base") positions (provided with grangesData)
and feature positions provided with grangesAnnotationsPos.
"Mod": the base modified. "Base": the base letter of the modified base.
Example: for Mod="6mA", Base="A"; for Mod="5mC", Base="C".
}
\keyword{internal}
