% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloModAn.R
\name{.IncludeModPosInMot}
\alias{.IncludeModPosInMot}
\title{IncludeModPosInMot Function (GloModAn)}
\usage{
.IncludeModPosInMot(
  nParamByMotif,
  cBaseLetterForMod,
  cModNameInOutput,
  nModPositionInMotif = 1,
  lExtractOnlyMotifWithMod = TRUE
)
}
\arguments{
\item{nParamByMotif}{A numeric vector named with motifs.}

\item{cBaseLetterForMod}{The name of the base letter of the modified base.}

\item{cModNameInOutput}{Name for the modification in the output.}

\item{nModPositionInMotif}{Number representing the position of the modification in the ranges
(after resizing if some base pairs were added). Defaults to 1.}

\item{lExtractOnlyMotifWithMod}{If TRUE, the numeric vector will be returned with only the
motifs that include DNA modifications. Defaults to TRUE.}
}
\description{
Return the numeric vector provided with new names to include the position of the modification.
}
\keyword{internal}
