% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloModAn.R
\name{.rescale}
\alias{.rescale}
\title{.rescale Function from seqLogo package (GloModAn)}
\usage{
.rescale(x, to = c(0.025, 0.975))
}
\description{
Internal function from seqLogo package. See seqLogo documentation for more details.
User should use the .rescale function from seqLogo and should not use directly this version.
}
\keyword{internal}
