% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloModAn.R
\name{DrawBarplotBothStrands}
\alias{DrawBarplotBothStrands}
\title{DrawBarplotBothStrands Function (GloModAn)}
\usage{
DrawBarplotBothStrands(
  nParamByContigForward,
  nParamByContigReverse,
  cContigNames,
  cGraphName,
  lIsOrderedLargestToSmallest = TRUE
)
}
\arguments{
\item{nParamByContigForward}{A numeric vector containing the parameter values of the forward strand to be plotted.
Must have the same order and same length as cContigNames and nParamByContigReverse.}

\item{nParamByContigReverse}{A numeric vector containing the parameter values of the reverse strand to be plotted.
Must have the same order and same length as nParamByContigForward and cContigNames.}

\item{cContigNames}{A character vector containing the names of the contigs.
Must have the same order and same length as nParamByContigForward and nParamByContigReverse.}

\item{cGraphName}{The graph name to be displayed on top of the plot.}

\item{lIsOrderedLargestToSmallest}{If contigs are ordered from largest to smallest contig,
add TRUE to display "(from largest to smallest)" below the plot. Defaults to FALSE.}
}
\description{
Return a barplot describing some parameter values provided by strand for each contig.
}
\examples{
DrawBarplotBothStrands(
  nParamByContigForward = c(100, 86, 75, 56),
  nParamByContigReverse = c(96, 88, 80, 83),
  cContigNames = c("chrI", "chrII", "chrIII", "chrIV"),
  cGraphName = "Mean Coverage per contig",
  lIsOrderedLargestToSmallest = TRUE
)
}
\keyword{DrawBarplotBothStrands}
