% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeQual.R
\name{DrawContigCumulLength}
\alias{DrawContigCumulLength}
\title{DrawContigCumulLength Function (SeQual)}
\usage{
DrawContigCumulLength(
  nContigCumsumLength,
  cOrgAssemblyName,
  lGridInBackground = FALSE,
  cLineColor = "red",
  nLineWidth = 2,
  nLineType = 1
)
}
\arguments{
\item{nContigCumsumLength}{A numeric vector containing the cumulative length of contigs (ordered from largest to smallest contig).}

\item{cOrgAssemblyName}{The name of the genome assembly corresponding to these contigs.}

\item{lGridInBackground}{If TRUE, add a grid in the background. Defaults to FALSE.}

\item{cLineColor}{The color of the line. Defaults to "red".}

\item{nLineWidth}{The width of the line. Defaults to 2.}

\item{nLineType}{The type of the line. Defaults to 1.}
}
\description{
Return a line-plot describing the cumulative length of contigs (ordered from largest to smallest contig).
}
\examples{
myGenome <- Biostrings::readDNAStringSet(system.file(
  package = "DNAModAnnot", "extdata",
  "ptetraurelia_mac_51_sca171819.fa"
))

myContig_cumul_len_t <- GetContigCumulLength(dnastringsetGenome = myGenome)

DrawContigCumulLength(
  nContigCumsumLength = myContig_cumul_len_t$cumsum_Mbp_length,
  cOrgAssemblyName = "ptetraurelia_mac_51"
)
}
\keyword{DrawContigCumulLength}
