% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloModAn.R
\name{DrawDistriHistBox}
\alias{DrawDistriHistBox}
\title{DrawDistriHistBox Function (GloModAn)}
\usage{
DrawDistriHistBox(
  nParam,
  cGraphName,
  cParamName,
  lTrimOutliers = FALSE,
  nXLimits = NULL
)
}
\arguments{
\item{nParam}{A numeric vector containing the parameter values to be plotted.}

\item{cGraphName}{The graph name to be displayed on top of the plot.}

\item{cParamName}{The name of the parameter to be displayed below the x-axis.}

\item{lTrimOutliers}{If TRUE, remove the outliers from the boxplot and trim the histogram to the borders
of the boxplot. Defaults to FALSE.}

\item{nXLimits}{A numeric vector giving the limits of the plot on the x-axis. If NULL, the limits will be set
to the minimum and the maximum of the nParam data (or to the inner fences (1.5*Interquartile Range) of the
boxplot if lTrimmingOutliers is TRUE). Defaults to NULL.}
}
\description{
Return a line-plot describing the cumulative length of contigs (ordered from largest to smallest contig).
}
\examples{
#loading genome
myGenome <- Biostrings::readDNAStringSet(system.file(package="DNAModAnnot", "extdata",
                                         "ptetraurelia_mac_51_sca171819.fa"))

#Preparing a gposPacBioCSV dataset
myGposPacBioCSV <-
   ImportPacBioCSV(cPacBioCSVPath = system.file(package="DNAModAnnot", "extdata",
                                 "ptetraurelia.bases.sca171819.csv"),
                   cSelectColumnsToExtract = c("refName", "tpl", "strand", "base",
                                               "score", "ipdRatio", "coverage"),
                   lKeepExtraColumnsInGPos = TRUE, lSortGPos = TRUE,
                   cContigToBeAnalyzed = names(myGenome))

DrawDistriHistBox(head(myGposPacBioCSV$coverage, 100000),
                         cGraphName = "Coverage distribution of some bases sequenced",
                         cParamName = "Coverage",
                         lTrimOutliers = TRUE )
}
\keyword{DrawDistriHistBox}
