% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModAnnot.R
\name{DrawModBaseCountsWithinFeature}
\alias{DrawModBaseCountsWithinFeature}
\title{DrawModBaseCountsWithinFeature Function (ModAnnot)}
\usage{
DrawModBaseCountsWithinFeature(
  grangesAnnotationsWithCountsByWindow,
  cFeatureName,
  cBaseMotif,
  cModMotif
)
}
\arguments{
\item{grangesAnnotationsWithCountsByWindow}{A GRanges object containing feature annotation with the counts:
\itemize{
  \item Modcount:     The   number of "Mod"  within this window  of this feature.
  \item ModcountSum:  Total number of "Mod"  within all  windows of this feature.
  \item Modprop:      (Modcount / ModcountSum)*100
  \item Basecount:    The   number of "Base" within this window  of this feature.
  \item BasecountSum: Total number of "Base" within all  windows of this feature.
  \item Baseprop:     (Basecount/BasecountSum)*100
}
The Genomic features categories must be in a column named "type".}

\item{cFeatureName}{Name of the feature (which is cut into windows) to be displayed.}

\item{cBaseMotif}{The name of the motif with the base letter of the modified base.}

\item{cModMotif}{The name of the motif with the modification in the output.}
}
\description{
Return a plot describing the proportion of the base modified (Mod) and the base letter
of the modified base (Base) between windows (of relative size) of feature provided.
Example: for Mod="6mA", Base="A"; for Mod="5mC", Base="C".
}
\examples{
# loading genome
myGenome <- Biostrings::readDNAStringSet(system.file(
  package = "DNAModAnnot", "extdata",
  "ptetraurelia_mac_51_sca171819.fa"
))

# loading annotation
library(rtracklayer)
myAnnotations <- readGFFAsGRanges(system.file(
  package = "DNAModAnnot", "extdata",
  "ptetraurelia_mac_51_annotation_v2.0_sca171819.gff3"
))

# Preparing a grangesPacBioGFF and a grangesPacBioCSV datasets
myGrangesPacBioGFF <-
  ImportPacBioGFF(
    cPacBioGFFPath = system.file(
      package = "DNAModAnnot", "extdata",
      "ptetraurelia.modifications.sca171819.gff"
    ),
    cNameModToExtract = "m6A",
    cModNameInOutput = "6mA",
    cContigToBeAnalyzed = names(myGenome)
  )
myGposPacBioCSV <-
  ImportPacBioCSV(
    cPacBioCSVPath = system.file(
      package = "DNAModAnnot", "extdata",
      "ptetraurelia.bases.sca171819.csv"
    ),
    cSelectColumnsToExtract = c(
      "refName", "tpl", "strand", "base",
      "score", "ipdRatio", "coverage"
    ),
    lKeepExtraColumnsInGPos = TRUE, lSortGPos = TRUE,
    cContigToBeAnalyzed = names(myGenome)
  )
myGposPacBioCSV <- myGposPacBioCSV[myGposPacBioCSV$base == "A"]

# Retrieve annotations with "Mod" and "Base" counts (and counts per kbp)
myAnn_ModBase_counts_by_window <-
  GetModBaseCountsWithinFeature(
    grangesAnnotations = myAnnotations[myAnnotations$type == "gene", ],
    grangesModPos = myGrangesPacBioGFF,
    gposModTargetBasePos = myGposPacBioCSV,
    nWindowsNb = 20
  )

DrawModBaseCountsWithinFeature(myAnn_ModBase_counts_by_window,
  cFeatureName = "gene",
  cBaseMotif = "A",
  cModMotif = "6mA"
)
}
\keyword{DrawModBaseCountsWithinFeature}
