% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModAnnot.R
\name{DrawModBasePropByFeature}
\alias{DrawModBasePropByFeature}
\title{DrawModBasePropByFeature Function (ModAnnot)}
\usage{
DrawModBasePropByFeature(
  grangesAnnotationsWithCounts,
  cFeaturesToCompare = c("gene", "intergenic"),
  lUseCountsPerkbp = FALSE,
  cBaseMotif,
  cModMotif
)
}
\arguments{
\item{grangesAnnotationsWithCounts}{A GRanges object based on grangesAnnotations with the counts:
\itemize{
  \item Modcount:           The number of "Mod"  within this feature.
  \item Modcount_perkbp:   (The number of "Mod"  within this feature divided by the size of the feature)*1000.
  \item Basecount:          The number of "Base" within this feature.
  \item Basecount_perkbp:  (The number of "Base" within this feature divided by the size of the feature)*1000.
}
The Genomic features categories must be in a column named "type".}

\item{cFeaturesToCompare}{Names of the features to be displayed and compared. Defaults to c("gene", "intergenic").}

\item{lUseCountsPerkbp}{If TRUE, use counts per kbp (Modcount_perkbp and Basecount_perkbp) to calculate
the proportion between features displayed. If FALSE, use counts (Modcount and Basecount) to calculate
the proportion between features displayed. Defaults to FALSE.}

\item{cBaseMotif}{The name of the motif with the base letter of the modified base.}

\item{cModMotif}{The name of the motif with the modification in the output.}
}
\description{
Return a plot describing, between the features to be displayed, the proportion
(or proportion per KiloBase pairs (kbp)) of
the base modified (Mod) and the base letter of the modified base (Base).
Example: for Mod="6mA", Base="A"; for Mod="5mC", Base="C".
}
\examples{
#loading genome
myGenome <- Biostrings::readDNAStringSet(system.file(package="DNAModAnnot", "extdata",
                                         "ptetraurelia_mac_51_sca171819.fa"))
myGrangesGenome <- GetGenomeGRanges(myGenome)

#loading annotation
library(rtracklayer)
myAnnotations <- readGFFAsGRanges(system.file(package="DNAModAnnot", "extdata",
                                          "ptetraurelia_mac_51_annotation_v2.0_sca171819.gff3"))
myAnnotations <- PredictMissingAnnotation(grangesAnnotations = myAnnotations,
                                          grangesGenome = myGrangesGenome,
                                          cFeaturesColName = "type",
                                          cGeneCategories = c("gene"),
                                          lAddIntronRangesUsingExon = TRUE)

#Preparing a grangesPacBioGFF and a grangesPacBioCSV datasets
myGrangesPacBioGFF <-
   ImportPacBioGFF(cPacBioGFFPath = system.file(package="DNAModAnnot", "extdata",
                                 "ptetraurelia.combinationC_merged_sca171819_sample.gff"),
                   cNameModToExtract = "m6A",
                   cModNameInOutput = "6mA",
                   cContigToBeAnalyzed = names(myGenome))
myGposPacBioCSV <-
   ImportPacBioCSV(cPacBioCSVPath = system.file(package="DNAModAnnot", "extdata",
                                 "ptetraurelia.combinationC.all.corrected_sca171819_sample.csv"),
                   cSelectColumnsToExtract = c("refName", "tpl", "strand", "base",
                                               "score", "ipdRatio", "coverage"),
                   lKeepExtraColumnsInGPos = TRUE, lSortGPos = TRUE,
                   cContigToBeAnalyzed = names(myGenome))
myGposPacBioCSV <- myGposPacBioCSV[myGposPacBioCSV$base == "A"]

#Retrieve annotations with "Mod" and "Base" counts (and counts per kbp)
myAnn_ModBase_counts <- GetModBaseCountsByFeature(grangesAnnotations=myAnnotations,
                                                  grangesModPos=myGrangesPacBioGFF,
                                                  gposModTargetBasePos=myGposPacBioCSV)

DrawModBasePropByFeature(grangesAnnotationsWithCounts=myAnn_ModBase_counts,
                                cFeaturesToCompare=c("gene", "intergenic"),
                                lUseCountsPerkbp=TRUE,
                                cBaseMotif="A",
                                cModMotif="6mA")
}
\keyword{DrawModBasePropByFeature}
