% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloModAn.R
\name{ExtractListModPosByModMotif}
\alias{ExtractListModPosByModMotif}
\title{ExtractListModPosByModMotif Function (GloModAn)}
\usage{
ExtractListModPosByModMotif(
  grangesModPos,
  grangesGenome,
  dnastringsetGenome,
  nUpstreamBpToAdd = 0,
  nDownstreamBpToAdd = 1,
  nModMotifMinProp,
  nModPositionInMotif = 1 + nUpstreamBpToAdd,
  cBaseLetterForMod,
  cModNameInOutput
)
}
\arguments{
\item{grangesModPos}{A GRanges object containing Modifications Positions data to be extracted with the sequence.}

\item{grangesGenome}{A GRanges object containing the width of each contig.}

\item{dnastringsetGenome}{A DNAStringSet object containing the sequence for each contig.}

\item{nUpstreamBpToAdd}{Number of base pairs to add upstream of the range from the GRanges object provided
to obtain some sequence upstream of range. If some new ranges do not fit in the ranges of the contigs (provided with grangesGenome),
those new ranges will be removed. New windows with gaps are also removed. Defaults to 0.}

\item{nDownstreamBpToAdd}{Number of base pairs to add downstream of the range from the GRanges object provided
to obtain some sequence downstream of range. If some new ranges do not fit in the ranges of the contigs (provided with grangesGenome),
those new ranges will be removed. New windows with gaps are also removed. Defaults to 0.}

\item{nModMotifMinProp}{A number indicating the false discovery rate to be used for filtering: this will allow to choose the closest threshold
below this number. Defaults to 0.05 (so fdr of 5\%).}

\item{nModPositionInMotif}{The position of the modification in the window after resizing with nUpstreamBpToAdd and nDownstreamBpToAdd.
If GRanges are 1-bp positions, then 1+nUpstreamBpToAdd will return the right position of the modification.
Defaults to 1+nUpstreamBpToAdd.}

\item{cBaseLetterForMod}{The name of the base letter of the modified base.}

\item{cModNameInOutput}{Name for the modification in the output.}
}
\value{
A list of 4 objects:
\describe{
  \item{motifs_to_analyse}{A character vector containing the sequence of motifs associated to the modification.}
  \item{mod_motif}{A character vector containing the sequence of motifs associated to the modification with the
  modification represented inside those motifs.}
  \item{motif_pct}{A table containing the percentage of modifications in each motif tested.}
  \item{GRangesbyMotif}{A list of GRanges objects with the sequence: one GRanges object by motif associated to the modification.}
}
}
\description{
Return the GRanges object provided with the sequence associated to each position (and can also retrieve the sequence around each position).
}
\examples{
#loading genome
myGenome <- Biostrings::readDNAStringSet(system.file(package="DNAModAnnot", "extdata",
                                         "ptetraurelia_mac_51_sca171819.fa"))
myGrangesGenome <- GetGenomeGRanges(myGenome)

#Preparing a grangesPacBioGFF dataset
myGrangesPacBioGFF <-
   ImportPacBioGFF(cPacBioGFFPath = system.file(package="DNAModAnnot", "extdata",
                                 "ptetraurelia.combinationC_merged_sca171819_sample.gff"),
                   cNameModToExtract = "m6A",
                   cModNameInOutput = "6mA",
                   cContigToBeAnalyzed = names(myGenome))

#Retrieve GRanges with sequence
myMotif_pct_and_GRangesList <- ExtractListModPosByModMotif(grangesModPos=myGrangesPacBioGFF,
                                                        grangesGenome = myGrangesGenome,
                                                        dnastringsetGenome = myGenome,
                                                        nUpstreamBpToAdd = 0,
                                                        nDownstreamBpToAdd = 1,
                                                        nModMotifMinProp = 0.05,
                                                        cBaseLetterForMod = "A",
                                                        cModNameInOutput = "6mA")

myMotif_pct_and_GRangesList$motifs_to_analyse
myMotif_pct_and_GRangesList$mod_motif
myMotif_pct_and_GRangesList$motif_pct
myMotif_pct_and_GRangesList$GRangesbyMotif
}
\keyword{ExtractListModPosByModMotif}
