% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeQual.R
\name{GetAssemblyReport}
\alias{GetAssemblyReport}
\title{GetAssemblyReport Function (SeQual)
importFrom base data.frame max sum return
importFrom Biostrings alphabetFrequency letterFrequency}
\usage{
GetAssemblyReport(dnastringsetGenome, cOrgAssemblyName)
}
\arguments{
\item{dnastringsetGenome}{A DNAStringSet object containing the sequence for each contig.}

\item{cOrgAssemblyName}{The name of the genome assembly provided.}
}
\description{
Return a report with global characteristics of genome assembly.
}
\examples{
myGenome <- Biostrings::readDNAStringSet(system.file(package="DNAModAnnot", "extdata",
                                         "ptetraurelia_mac_51_sca171819.fa"))

myReport <- GetAssemblyReport(dnastringsetGenome = myGenome,
                                  cOrgAssemblyName = "ptetraurelia_mac_51")

myReport
}
\keyword{GetAssemblyReport}
