% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaLoad.R
\name{GetGenomeGRanges}
\alias{GetGenomeGRanges}
\title{GetGenomeGRanges Function (DaLoad)}
\usage{
GetGenomeGRanges(dnastringsetGenome)
}
\arguments{
\item{dnastringsetGenome}{A DNAStringSet object to convert as a GRanges object.
If sequences have no names, these will be named with "seq" and a number
(corresponding to the order in the DNAStringSet object) in the GRanges object.}
}
\description{
Return a GRanges object from a DNAStringSet object with ranges from 1 to the sequences width (for both strands).
}
\examples{
mySeqs <- Biostrings::DNAStringSet(c("ACCATTGATTAT", "AATATCGACTA", "GACTAT"))
myRanges <- GetGenomeGRanges(mySeqs)
myRanges
}
\keyword{GetGenomeGRanges}
