% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaLoad.R
\name{GetGposCenterFromGRanges}
\alias{GetGposCenterFromGRanges}
\title{GetGposCenterFromGRanges Function (DaLoad)}
\usage{
GetGposCenterFromGRanges(grangesData)
}
\arguments{
\item{grangesData}{A GRanges object to be converted as a GPos object. The center of each range will be processed in the GPos object output.}
}
\description{
Retrieve, in a GPos object, the positions of the center of the ranges from a GRanges object.
}
\examples{
myRanges <- as(c("chrI:300-500:+", "chrI:308-680:+" , "chrII:30-550:-"), "GRanges")
myCenterPos <- GetGposCenterFromGRanges(grangesData=myRanges)
myCenterPos
}
\keyword{GetGposCenterFromGRanges}
