% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloModAn.R
\name{GetGRangesWindowSeqandParam}
\alias{GetGRangesWindowSeqandParam}
\title{GetGRangesWindowSeqandParam Function (GloModAn)}
\usage{
GetGRangesWindowSeqandParam(
  grangesData,
  grangesGenome,
  dnastringsetGenome,
  nUpstreamBpToAdd = 0,
  nDownstreamBpToAdd = 0
)
}
\arguments{
\item{grangesData}{A GRanges object containing Modifications Positions data to be extracted with the sequence.}

\item{grangesGenome}{A GRanges object containing the width of each contig.}

\item{dnastringsetGenome}{A DNAStringSet object containing the sequence for each contig.}

\item{nUpstreamBpToAdd}{Number of base pairs to add upstream of the range from the GRanges object provided
to obtain some sequence upstream of range. If some new ranges do not fit in the ranges of the contigs (provided with grangesGenome),
those new ranges will be removed. New windows with gaps are also removed. Defaults to 0.}

\item{nDownstreamBpToAdd}{Number of base pairs to add downstream of the range from the GRanges object provided
to obtain some sequence downstream of range. If some new ranges do not fit in the ranges of the contigs (provided with grangesGenome),
those new ranges will be removed. New windows with gaps are also removed. Defaults to 0.}
}
\description{
Return the GRanges object provided with the sequence associated to each position (and can also retrieve the sequence around each position).
}
\examples{
#loading genome
myGenome <- Biostrings::readDNAStringSet(system.file(package="DNAModAnnot", "extdata",
                                         "ptetraurelia_mac_51_sca171819.fa"))
myGrangesGenome <- GetGenomeGRanges(myGenome)

#Preparing a grangesPacBioGFF datasets
myGrangesPacBioGFF <-
   ImportPacBioGFF(cPacBioGFFPath = system.file(package="DNAModAnnot", "extdata",
                                 "ptetraurelia.modifications.sca171819.gff"),
                   cNameModToExtract = "m6A",
                   cModNameInOutput = "6mA",
                   cContigToBeAnalyzed = names(myGenome))

#Retrieve GRanges with sequence
myPositions_Mod_Granges_wSeq <- GetGRangesWindowSeqandParam(grangesData=myGrangesPacBioGFF,
                                                            grangesGenome = myGrangesGenome,
                                                            dnastringsetGenome = myGenome,
                                                            nUpstreamBpToAdd = 5,
                                                            nDownstreamBpToAdd = 5)
myPositions_Mod_Granges_wSeq
}
\keyword{GetGRangesWindowSeqandParam}
