% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloModAn.R
\name{GetMeanParamByContig}
\alias{GetMeanParamByContig}
\title{GetMeanParamByContig Function (GloModAn)}
\usage{
GetMeanParamByContig(grangesData, dnastringsetGenome, cParamName)
}
\arguments{
\item{grangesData}{A GRanges-like object containing, in the extra columns, the parameter to be analysed.}

\item{dnastringsetGenome}{A DNAStringSet object containing the sequence for each contig.}

\item{cParamName}{The name of the column containing the parameter to be analysed.}
}
\value{
A list composed of 3 dataframes: 1 dataframe for each strand and 1 dataframe with both strands. In each dataframe:
\itemize{
  \item refName: The names of each contig.
  \item strand: The strand of each contig.
  \item width: The width of each contig.
  \item nb_sequenced: The number of bases sequenced by strand for each contig.
  \item seqPct: The percentage of bases sequenced by strand for each contig (percentage of sequencing).
}
}
\description{
Return a list with the mean by strand of the parameter provided for all scaffolds of genome assembly provided.
}
\examples{
myGenome <- Biostrings::readDNAStringSet(system.file(
  package = "DNAModAnnot", "extdata",
  "ptetraurelia_mac_51_sca171819.fa"
))

# Preparing a gposPacBioCSV dataset
myGposPacBioCSV <-
  ImportPacBioCSV(
    cPacBioCSVPath = system.file(
      package = "DNAModAnnot", "extdata",
      "ptetraurelia.bases.sca171819.csv"
    ),
    cSelectColumnsToExtract = c(
      "refName", "tpl", "strand", "base",
      "score", "ipdRatio", "coverage"
    ),
    lKeepExtraColumnsInGPos = TRUE, lSortGPos = TRUE,
    cContigToBeAnalyzed = names(myGenome)
  )

myMean_cov_list <- GetMeanParamByContig(
  grangesData = myGposPacBioCSV,
  dnastringsetGenome = myGenome,
  cParamName = "coverage"
)
myMean_cov_list
}
\keyword{GetMeanParamByContig}
