% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaLoad.R
\name{ImportDeepSignalModFrequency}
\alias{ImportDeepSignalModFrequency}
\title{ImportDeepSignalModFrequency Function (DaLoad)}
\usage{
ImportDeepSignalModFrequency(
  cDeepSignalModPath,
  cColumnNames = c("chrom", "pos", "strand", "pos_in_strand", "prob_0_sum",
    "prob_1_sum", "count_modified", "count_unmodified", "coverage",
    "modification_frequency", "k_mer"),
  cSelectColumnsToExtract = c("chrom", "pos", "strand", "prob_0_sum", "prob_1_sum",
    "count_modified", "count_unmodified", "coverage", "modification_frequency", "k_mer"),
  lSortGPos = TRUE,
  cContigToBeAnalyzed,
  lKeepSequence = TRUE
)
}
\arguments{
\item{cDeepSignalModPath}{Path to a DeepSignal call_modification_frequency.py output file containing data from all target sites.}

\item{cColumnNames}{Names for each column in the DeepSignal call_modification_frequency.py output file.
Should not be changed unless some columns are missing in the file to be imported.
Defaults to c("chrom", "pos", "strand", "pos_in_strand",
"prob_0_sum", "prob_1_sum", "count_modified", "count_unmodified",
"coverage", "modification_frequency", "k_mer")}

\item{cSelectColumnsToExtract}{Names of columns to extract from DeepSignal call_modification_frequency.py output file.
Less there are columns, faster the file will be loaded.
The columns "chrom", "pos" and "strand" are mandatory to convert to a GPos object.
Defaults to c("chrom", "pos", "strand", "prob_0_sum", "prob_1_sum", "count_modified", "count_unmodified",
"coverage", "modification_frequency", "k_mer")}

\item{lSortGPos}{If TRUE, the GPos object will be sorted before being returned: the function will take a longer time
to proceed but the GPos Object will require less memory.}

\item{cContigToBeAnalyzed}{Names of contigs for which the data will be kept.
If NULL, data from all contigs available will be imported. Defaults to NULL.}

\item{lKeepSequence}{If TRUE, the sequence of the base will be retained in one column.
Otherwise, it will be discarded to reduce object size. Defaults to TRUE.}
}
\description{
Import DeepSignal call_modification_frequency.py output file and convert it as an UnStitched GPos object.
}
\examples{
#Loading Nanopore data
myDeepSignalModPath <- system.file(
  package="DNAModAnnot", "extdata",
  "FAB39088-288418386-Chr1.CpG.call_mods.frequency.tsv")
mygposDeepSignalModBase <- ImportDeepSignalModFrequency(cDeepSignalModPath=myDeepSignalModPath,
                                                        lSortGPos=TRUE,
                                                        cContigToBeAnalyzed = "all")
mygposDeepSignalModBase
}
\keyword{ImportDeepSignalModFrequency}
