% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloModAn.R
\name{letterA}
\alias{letterA}
\title{letterA Function from seqLogo package (GloModAn)}
\usage{
letterA(x.pos, y.pos, ht, wt, fill = "#61D04F", id = NULL)
}
\description{
Internal function from seqLogo package. See seqLogo documentation for more details.
User should use the letterA function from seqLogo and should not use directly this version.
}
\keyword{internal}
