% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloModAn.R
\name{letterC}
\alias{letterC}
\title{letterC Function from seqLogo package (GloModAn)}
\usage{
letterC(x.pos, y.pos, ht, wt, fill = "#2297E6", id = NULL)
}
\description{
Internal function from seqLogo package. See seqLogo documentation for more details.
User should use the letterC function from seqLogo and should not use directly this version.
}
\keyword{internal}
