% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloModAn.R
\name{letterG}
\alias{letterG}
\title{letterG Function from seqLogo package (GloModAn)}
\usage{
letterG(x.pos, y.pos, ht, wt, fill = "#F5C710", id = NULL)
}
\description{
Internal function from seqLogo package. See seqLogo documentation for more details.
User should use the letterG function from seqLogo and should not use directly this version.
}
\keyword{internal}
