% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloModAn.R
\name{letterU}
\alias{letterU}
\title{letterU Function from seqLogo package (GloModAn)}
\usage{
letterU(x.pos, y.pos, ht, wt, fill = "#DF536B", id = NULL)
}
\description{
Internal function from seqLogo package. See seqLogo documentation for more details.
User should use the letterU function from seqLogo and should not use directly this version.
}
\keyword{internal}
